requirements.txt
homekit[IP]
questionary
rich
python-dotenv